Actually become python 2 and 3 compatible
Turns out our first pass at running the tests via tox
would *load* the tests but not actuall run them.  So
we had fixed the import error but in execution it still
broke in python 3.

Here we upgraded the Dockerfile and tox to use python 3.7
and got the test passing.

Co-authored-by: Chris Cenatiempo <chrisc@realgeeks.com>